+++ ChangeLog.md
@@ -2,12 +2,12 @@
 
 ## ChangeLog
 
-### 1.0.0-rc
+### 1.0.0
 
-    Release date: 
+    Release date: 2025- 6-30
 
     - Initial release
 
 ---
 P. Wieser
-- Last updated on 2024, Jul. 29th
+- Last updated on 2025, Jun. 30th